+++ src/main/java/org/sopt/domain/post/repository/PostRepository.java
@@ -1,8 +1,8 @@
-package org.sopt.repository;
+package org.sopt.domain.post.repository;
 
 import java.util.List;
 import java.util.Optional;
-import org.sopt.domain.Post;
+import org.sopt.domain.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Repository;
