+++ src/main/java/org/sopt/domain/post/validator/PostValidationRule.java
@@ -1,4 +1,4 @@
-package org.sopt.validator;
+package org.sopt.domain.post.validator;
 
 public interface PostValidationRule {
     void validate(String title);
